MNS 微调

+++ framework/util/MNS.class.php
@@ -17,15 +17,15 @@ define("PITHY_CONFIG_MNS", serialize(
     array(
          "MNS" => array(
             "topic" => array(
-                "key" => "$key",
-                "host" => "$accountId.mns.cn-shenzhen.aliyuncs.com",
                 "type" => "topic",
+                "host" => "$accountId.mns.cn-shenzhen.aliyuncs.com",
+                "key" => "$key",
                 
             ),
-            "query" => array(
-                "key" => "$key",
+            "queue" => array(
+                "type" => "queue",
                 "host" => "$accountId.mns.cn-shenzhen.aliyuncs.com",
-                "type" => "query",
+                "key" => "$key",
             ),
         ), 
     )
@@ -51,8 +51,11 @@ class MNS  {
         if (empty($config)) {
             return trigger_error("MNS 配置有误", E_USER_ERROR);
         }        
-        if (empty($name) || !is_string($name) || !isset($config[$name],$config[$name]["key"])) {
+        if (empty($name) || !is_string($name) || !isset($config[$name], $config[$name]["key"], $config[$name]["type"])) {
             return trigger_error("MNS 初始化出错", E_USER_ERROR);
+        }        
+        if (!in_array($config[$name]["type"], array("topic", "queue"))) {
+            return trigger_error("MNS 消息类型出错", E_USER_ERROR);
         }
         
         $this->name = $name;
@@ -62,6 +65,7 @@ class MNS  {
     
     // Topic 订阅
     public function subscribe($name, $tag, $point){
+        if ($this->check("topic")) return false;
         $retry = substr($point, 0, 4) == "http" ? "EXPONENTIAL_DECAY_RETRY" : "BACKOFF_RETRY";
         $format = substr($point, 0, 4) == "http" ? "SIMPLIFIED" : "JSON";
         $content = "<?xml version='1.0' encoding='utf-8'?>\n<Subscription xmlns='http://mns.aliyuncs.com/doc/v1/'>\n<FilterTag>{$tag}</FilterTag>\n<Endpoint>{$point}</Endpoint>\n<NotifyStrategy>{$retry}</NotifyStrategy>\n<NotifyContentFormat>{$format}</NotifyContentFormat>\n</Subscription>";
@@ -71,12 +75,14 @@ class MNS  {
     
     // Topic 取消订阅
     public function unsubscribe($name){
+        if ($this->check("topic")) return false;
         $rtn = $this->request("DELETE", "/topics/{$this->name}/subscriptions/{$name}");
         return $rtn == 204;
     }
     
     // Topic 发布消息
     public function publish($body, $tag, $attributes = ""){
+        if ($this->check("topic")) return false;
         $content = "<?xml version='1.0' encoding='utf-8'?>\n<Message xmlns='http://mns.aliyuncs.com/doc/v1/'>\n<MessageBody>".base64_encode(json_encode($body))."</MessageBody>\n<MessageTag>{$tag}</MessageTag>\n{$attributes}</Message>";
         $rtn = $this->request("POST", "/topics/{$this->name}/messages", $content);
         return $rtn == 201;
@@ -85,6 +91,7 @@ class MNS  {
     
     // 发送消息
     public function send($body, $delay = 0){
+        if ($this->check("queue")) return false;
         $content = "<?xml version='1.0' encoding='UTF-8'?>\n<Message xmlns='http://mns.aliyuncs.com/doc/v1/'>\n<MessageBody>".base64_encode(json_encode($body))."</MessageBody>\n<DelaySeconds>".intval($delay)."</DelaySeconds>\n</Message>";
         $rtn = $this->request("POST", "/queues/{$this->name}/messages", $content);
         return $rtn == 201;
@@ -92,6 +99,7 @@ class MNS  {
     
     // 接收消息
     public function receive($destroy = true){
+        if ($this->check("queue")) return false;
         $rtn = $this->request("GET", "/queues/{$this->name}/messages");
         $rtn == 200 && $destroy && $this->destroy();
         return $rtn == 200 ? $this->message["MessageBody"] : null;
@@ -99,6 +107,7 @@ class MNS  {
     
     // 延长消息可见时间
     public function delay($timeout, $id = ""){
+        if ($this->check("queue")) return false;
         empty($id) && isset($this->message["ReceiptHandle"]) && $id = $this->message["ReceiptHandle"];
         $rtn = $this->request("PUT", "/queues/{$this->name}/messages?receiptHandle={$id}&visibilityTimeout=".intval($timeout));
         return $rtn == 200;
@@ -106,11 +115,20 @@ class MNS  {
     
     // 删除消息
     public function destroy($id = ""){
+        if ($this->check("queue")) return false;
         empty($id) && isset($this->message["ReceiptHandle"]) && $id = $this->message["ReceiptHandle"];
         $rtn = $this->request("DELETE", "/queues/{$this->name}/messages?receiptHandle={$id}");
         return $rtn == 204;
     }
     
+
+    protected function check($type){
+        if ($this->config["type"] != $type){
+            trigger_error("该消息的类型不支持此操作", E_USER_ERROR);
+            return false;
+        }
+        return true;
+    }
     
     protected function request($method, $uri, $content = ""){
